+++ astro.config.mjs
@@ -1,6 +1,5 @@
 import { defineConfig } from 'astro/config';
 import mdx from "@astrojs/mdx";
-
 import react from "@astrojs/react";
 
 // https://astro.build/config
