--- src/components/search/SearchBar.astro
@@ -1,7 +1,0 @@
----
-import Search from './Search.jsx';
-const allPosts = await Astro.glob('../../content/resources/*.{md,mdx}');
-const posts = allPosts.filter((post) => !post.frontmatter.draft && post.frontmatter.slug);
----
-
-<Search client:load searchList={posts} />
