--- src/pages/[slug].astro
@@ -1,23 +1,0 @@
----
-import Base from "../layouts/Base.astro";
-
-export async function getStaticPaths() {
-	const allPosts = await Astro.glob('../content/resources/*.{md,mdx}');
-	const posts = allPosts.filter((post) => !post.frontmatter.draft && post.frontmatter.slug);
-
-	return posts.map((post) => ({
-		params: {
-			slug: post.frontmatter.slug,
-		},
-		props: { post },
-	}));
-}
-
-const { post } = Astro.props;
----
-
-<Base title={post.frontmatter.title}>
-	<h2>{post.frontmatter.title}</h2>
-	<p>{post.frontmatter.description}</p>
-	<post.Content />
-</Base>
