+++ src/pages/resources.astro
@@ -2,9 +2,8 @@
 import { getCollection } from "astro:content";
 import { Navbar } from "../components/resources";
 import { TopicCard } from "../components/core";
-import { Icon } from "astro-icon";
+import { Search } from "../components/islands";
 import Base from "../layouts/Base.astro";
-import SearchBar from "../components/search/SearchBar.astro";
 import "./resources.style.scss";
 
 const metadata = {
@@ -15,8 +14,6 @@ const metadata = {
 };
 
 const resourcePages = await getCollection("resources");
-const allPosts = await Astro.glob('../content/resources/*.{md,mdx}');
-const posts = allPosts.filter((post) => !post.frontmatter.draft && post.frontmatter.slug);
 ---
 
 <Base {metadata}>
@@ -27,7 +24,7 @@ const posts = allPosts.filter((post) => !post.frontmatter.draft && post.frontmat
     <p class="resources-page__heading-subtitle">
       Resources created by students for students.
     </p>
-    <SearchBar searchList={posts} />
+    <Search client:load searchList={resourcePages} />
   </div>
   <div class="resources-page__cards">
     <TopicCard
